Massively restructure the project files
