+++ GW2LogProofs.vcxproj
@@ -11,57 +11,62 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="src\bosses.cpp" />
+    <ClCompile Include="src\core\bosses.cpp" />
     <ClCompile Include="src\entry.cpp" />
-    <ClCompile Include="src\gui.cpp" />
+    <ClCompile Include="src\ui\gui.cpp" />
     <ClCompile Include="src\imgui\imgui.cpp" />
     <ClCompile Include="src\imgui\imgui_demo.cpp" />
     <ClCompile Include="src\imgui\imgui_draw.cpp" />
     <ClCompile Include="src\imgui\imgui_tables.cpp" />
     <ClCompile Include="src\imgui\imgui_widgets.cpp" />
-    <ClCompile Include="src\log_proofs.cpp" />
-    <ClCompile Include="src\settings.cpp" />
-    <ClCompile Include="src\shared.cpp" />
+    <ClCompile Include="src\core\log_proofs.cpp" />
+    <ClCompile Include="src\core\settings.cpp" />
+    <ClCompile Include="src\core\shared.cpp" />
     <ClCompile Include="src\unofficial_extras\KeyBindHelper.cpp" />
     <ClCompile Include="src\unofficial_extras\KeyBindsTranslation.cpp" />
-    <ClCompile Include="src\kp_loader.cpp" />
-    <ClCompile Include="src\provider_registry.cpp" />
-    <ClCompile Include="src\ui_renderer.cpp" />
-    <ClCompile Include="src\gui_settings.cpp" />
-    <ClCompile Include="src\providers\wingman_provider.cpp" />
-    <ClCompile Include="src\providers\kpme_provider.cpp" />
+    <ClCompile Include="src\providers\common\provider_registry.cpp" />
+    <ClCompile Include="src\utils\httpclient.cpp" />
+    <ClCompile Include="src\ui\ui_renderer.cpp" />
+    <ClCompile Include="src\ui\gui_settings.cpp" />
+    <ClCompile Include="src\providers\wingman\wingman_provider.cpp" />
+    <ClCompile Include="src\providers\wingman\wingman_client.cpp" />
+    <ClCompile Include="src\providers\kpme\kpme_provider.cpp" />
+    <ClCompile Include="src\providers\kpme\kpme_client.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="src\imgui_extensions.h" />
+    <ClInclude Include="src\ui\imgui_extensions.h" />
     <ClInclude Include="src\arcdps\arcdps.h" />
-    <ClInclude Include="src\bosses.h" />
-    <ClInclude Include="src\gui.h" />
-    <ClInclude Include="src\httpclient.hpp" />
+    <ClInclude Include="src\core\bosses.h" />
+    <ClInclude Include="src\ui\gui.h" />
+    <ClInclude Include="src\utils\httpclient.h" />
     <ClInclude Include="src\imgui\imconfig.h" />
     <ClInclude Include="src\imgui\imgui.h" />
     <ClInclude Include="src\imgui\imgui_internal.h" />
     <ClInclude Include="src\imgui\imstb_rectpack.h" />
     <ClInclude Include="src\imgui\imstb_textedit.h" />
     <ClInclude Include="src\imgui\imstb_truetype.h" />
-    <ClInclude Include="src\log_proofs.h" />
+    <ClInclude Include="src\core\log_proofs.h" />
     <ClInclude Include="src\nexus\Nexus.h" />
     <ClInclude Include="src\nlohmann\json.hpp" />
     <ClInclude Include="src\resource.h" />
-    <ClInclude Include="src\settings.h" />
-    <ClInclude Include="src\shared.h" />
-    <ClInclude Include="src\threadpool.hpp" />
+    <ClInclude Include="src\core\settings.h" />
+    <ClInclude Include="src\core\shared.h" />
+    <ClInclude Include="src\utils\threadpool.hpp" />
     <ClInclude Include="src\unofficial_extras\Definitions.h" />
     <ClInclude Include="src\unofficial_extras\KeyBindHelper.h" />
     <ClInclude Include="src\unofficial_extras\KeyBindsTranslation.h" />
     <ClInclude Include="src\unofficial_extras\KeyBindStructs.h" />
     <ClInclude Include="src\version.h" />
-    <ClInclude Include="src\kp_loader.h" />
-    <ClInclude Include="src\data_provider.h" />
-    <ClInclude Include="src\provider_registry.h" />
-    <ClInclude Include="src\player_data.h" />
-    <ClInclude Include="src\ui_renderer.h" />
-    <ClInclude Include="src\providers\wingman_provider.h" />
-    <ClInclude Include="src\providers\kpme_provider.h" />
+    <ClInclude Include="src\providers\common\data_provider.h" />
+    <ClInclude Include="src\providers\common\provider_registry.h" />
+    <ClInclude Include="src\core\player_data.h" />
+    <ClInclude Include="src\ui\ui_renderer.h" />
+    <ClInclude Include="src\providers\wingman\wingman_provider.h" />
+    <ClInclude Include="src\providers\wingman\wingman_client.h" />
+    <ClInclude Include="src\providers\wingman\wingman_types.h" />
+    <ClInclude Include="src\providers\kpme\kpme_provider.h" />
+    <ClInclude Include="src\providers\kpme\kpme_client.h" />
+    <ClInclude Include="src\providers\kpme\kpme_types.h" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="src\resource.rc" />
