+++ src/core/log_proofs.h
@@ -6,12 +6,11 @@
 #include <string>
 #include <vector>
 
-#include "unofficial_extras/Definitions.h"
+#include "../unofficial_extras/Definitions.h"
 
-#include "kp_loader.h"
-#include "threadpool.hpp"
+#include "../utils/threadpool.hpp"
 #include "player_data.h"
-#include "provider_registry.h"
+#include "../providers/common/provider_registry.h"
 
 namespace LogProofs {
 	// Legacy compatibility - will be removed after full migration
