+++ src/providers/common/provider_registry.cpp
