+++ src/providers/common/provider_registry.h
