+++ src/providers/kpme/kpme_client.h
@@ -1,0 +1,13 @@
+#ifndef KPME_CLIENT_H
+#define KPME_CLIENT_H
+
+#include "kpme_types.h"
+
+namespace Kpme {
+    class KpmeClient {
+    public:
+        KpmeResponse GetKp(const std::string& account);
+    };
+}
+
+#endif
