+++ src/providers/wingman/wingman_provider.h
@@ -1,8 +1,8 @@
 #ifndef WINGMAN_PROVIDER_H
 #define WINGMAN_PROVIDER_H
 
-#include "../data_provider.h"
-#include "../kp_loader.h"
+#include "../common/data_provider.h"
+#include "wingman_client.h"
 
 class WingmanProvider : public IDataProvider {
 public:
@@ -12,6 +12,7 @@ public:
     std::vector<std::string> GetSupportedProofTypes() const override;
 
 private:
+    Wingman::WingmanClient client_;
     PlayerProofData ConvertWingmanResponse(const Wingman::WingmanResponse& response);
 };
